+++ dbms/src/Common/ThreadManager.cpp
@@ -16,6 +16,7 @@
 #include <Common/ThreadFactory.h>
 #include <Common/ThreadManager.h>
 #include <Common/wrapInvocable.h>
+#include <common/ThreadPool.h>
 
 namespace DB
 {
