+++ dbms/src/Storages/Transaction/LearnerRead.cpp
@@ -29,7 +29,6 @@
 #include <Storages/Transaction/TMTContext.h>
 #include <Storages/Transaction/Types.h>
 #include <Storages/Transaction/Utils.h>
-#include <common/ThreadPool.h>
 #include <common/likely.h>
 #include <common/logger_useful.h>
 #include <fmt/chrono.h>
